relocate tools
