+++ src/lib/features/ai/tools/index.ts
@@ -1,0 +1,9 @@
+export type { ContextType } from './types.js';
+export {
+	githubLinkHandler,
+	svelteDevLinkHandler,
+	fetchLinkContentTool
+} from './fetch-link-content.js';
+export { chatSearchTool } from './chat-search.js';
+export { getChat } from './get-chat.js';
+export { webSearch } from './web-search.js';
