+++ src/lib/features/ai/tools/types.ts
@@ -1,0 +1,10 @@
+import type { ActionCtx } from '../../../convex/_generated/server';
+import type { Id } from '../../../convex/_generated/dataModel';
+
+export type ContextType = {
+	env: {
+		GITHUB_TOKEN: string | undefined;
+	};
+	ctx: ActionCtx;
+	chatId: Id<'chats'>;
+};
